🐛 Favor `_ssi` field over `_tsi` for split from PDF
As a `_tsi` it's tokenized and we can't use it reliably for join
queries.

As an `_ssi` the following Solr `q` value will allow me to query all
documents that have an `application/pdf` mime_type (e.g. a FileSet) and
do not have child works that were split from the given PDF.

```
mime_type_ssi:application/pdf AND
-_query_:"{!v='id:split_from_pdf_id_tsi'}"
```

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/681