🎁 Add config for skipping splitting by file suffix
Prior to this commit, we always split every PDF file.  And for some
situations, we do not want split a PDF (e.g. In Adventist when we have a
work that attaches a reader and archival PDF; we don't need/want to
split the reader PDF, as we're splitting the archival PDF).