♻️ Extract direct ActiveFedora calls to adapter
With Valkyrie looming large, this change refactors the code towards an
adapter pattern; one that will allow for creating a Valkyrie adapter and
then configuring IIIFPrint accordingly.

We'll also need to consider how we duplicate the actor work within a
transaction and/or publisher/listener setup.

There are likely other places where we're making assumptions about
presenters and/or the data models we're processing.