♻️ Favor explicit method over class attribute
The `class_attribute` best serves cases where we might adjust
configuration at run-time or during initialization.  In the case of of
the class method I am removing, it's subtly different than the ideal use
case for `class_attribute`.

The difference is such:

We include the configuration module into a model; this is the run time
configuration.  And because we've "configured IIIF Print" for that
model, we always will have a `.iiif_print_config?` of true.