🎁 Add listener for handling file set attachment
Why a listener and not a transaction?  In part because the moment I want
to perform the conditional enqueuing is at the point where the
`Hyrax::WorkUploadsHandler` does it's job.  That is when we have:

- the parent work
- the file set
- the original file
- the user

The `Hyrax::WorkUploadsHandler` is most analogous to the behavior in
`Hyrax::Actors::FileSetActor#attach_to_work` and
`Hyrax::Actors::FileSetActor#create_content`.  Fortunately, Hyrax's
transaction and upload handler remove the conditional handling we needed
between uploading a remote file and directly uploading a file.

Related to:

- https://github.com/scientist-softserv/hykuup_knapsack/issues/35
- https://github.com/scientist-softserv/hykuup_knapsack/issues/99
- https://github.com/scientist-softserv/iiif_print/issues/312