🧹 Avoid mutating super's array
Prior to this commit, each time we called `config_search_paths` we would
append `IiifPrint::Engine.root` to the instances array.  This mean that
we could continue to prepend many times resulting in many entries of
`IiifPrint::Engine.root` in the instance's array.

With this change we avoid mutating the array in `super`.