♻️ Remove conditionals and account for Valkyrie
The `IiifPrint.model_configuration` is something we use to create a
module that we mixin to a model; either `ActiveFedor::Base` or
`Valkyrie::Resource`.

As such we know when the model is configured for IiifPrint.  It follows
that we can remove the `iiif_print_config?` methods.

Further, to account for Valkyrie indexers favoring the `#to_solr` method
and ActiveFedora indexers favoring `#generate_solr_document` we
introduce a bit of sleuthing to see which type we have.  But also
concede that if you've implemented either method we're going to
duplicate that behavior.

So we get to remove one module and two methods; and remove a potential
timing issue regarding when we mixing the model_configuration and when
we loop through the work_types to include additional logic.