🧹 Remove `ENV['HYKU_IIIF_PRINT']` conditional
The knowledge of why we do or don't have `HYKU_IIIF_PRINT` is a bit
odd. I'm not excited to have an `ENV` from Hyku show up here; especially
an `ENV` that only partially removes the full functionality of IIIF
Print.

What I'm wondering is if we should have Hyku decorate the
work_resource.set_child_flag to account for it's implementation foibles
relative to how Hyku is using the HYKU_IIIF_PRINT.

Put another, when we see this variable, my read is HYKU_IIIF_PRINT ==
false means don't do anything IIIF Print related. But that is not what
it means in Hyku, thus we should limit the knowledge/foibles of that ENV
to Hyku only.