🎁 Implement delete work for valyrie adapter
This commit will add an override for the `DeleteAllFileSets` step so we
can insert the `iiif_print_conditionally_destroy_spawned_children` with
args, we want to pass in the user so we can publish it to any relevant
listeners.  We also implement the valkyrie way of deleting split child
works in the `destroy_children_split_from` method.