🐛 Propogate permissions to child works & filesets
Copies the permissions from the parent work to each of the spawned child
works.

Prior to this commit, the child works split off of parent works were
assigned permissions according to the parent's current visibility (in
the case of a lease or embargo, this means they were assigned the value
during lease or embargo).

This would protect the child work during the lease or embargo period but
not allow it to be maintained by the expiry jobs, resulting in an
incorrect visibility setting at the end of the lease or embargo.