🐛 Handle blank query and snippets
This commit will address blank catalog searches and how it interacts
with snippets.  Prior, when the user submitted a blank catalog search,
the entire indexed text will be returned.  This is not desired because
if we're looking at a book with a hundred pages, then that's a lot of
rendered text.  Instead, if we don't have a query string then we will
also return.