Fix search & highlighting
Search was broken in cases where there is no json file to search (i.e.
the pdf file didn't have the json file). There may have been a change in
a blacklight_iiif_search version update, which made the INVALID text
that we used in this situation lock up the search.

This change returns the normal search behavior.